Minor bug fixes (#4)
* Make armor list and item list drag-and-droppable

To support creating something in an actor's inventory and saving it to the item sidebar/compendium later. Also supports any module interaction that requires drag-and-drop.

* Missing quotes in "templates" armor list

* Fix unequipped armor counting for damage resistance on character sheet

* Calculate actor.system.dr as part of prepareDerivedData

So we can put it on the character sheet

* Replace references to TokenDocument._actor with TokenDocument.actor

_actor property is deprecated in V10 and removed in V11. actor() accessor takes its place.

* Assign calculated DR back to actor for display

* Apply DR essence to actor DR

DR's "greater" flag isn't being saved by Handlebars when the essence is selected (or any essence that only exists in Greater form, for that matter).  Added some new logic to the magic item calculations to ensure it's set properly.

* Delete print statements from DR fix testing

* Set initial DR to zero during calculations

Fixes bug where a character's DR when the server closes becomes their "base" DR when reloading the server onto which armor et al. is added, effectively doubling DR unless manually corrected.

* Remove redundant NumericWeaponProperty for ArmorPenetration

Was causing both AP and Armor Penetration to be printed on weapon customization sheet, which was covering up options on the lower row.

* Give "general" items an armor-style header

So we can set the item picture.